Fix: Issue Item Link 태그 오류 수정
